* initialize XRay in base application class, and make Show notification action button configurable
