+++ android/app/src/main/AndroidManifest.xml
@@ -8,6 +8,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+        android:name="App"
         android:theme="@style/AppTheme">
         <activity android:name="com.applicaster.xray.example.ui.MainActivity">
             <intent-filter>
